+++ src/Rule/General/DisabledConcatenationWithTranslatedStringsRule.php
@@ -11,6 +11,7 @@ use PhpParser\Node\Expr\BinaryOp\Concat;
 use PhpParser\Node\Expr\FuncCall;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\StaticCall;
+use PhpParser\Node\Scalar\String_;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\Rule;
 use PHPStan\Rules\RuleErrorBuilder;
@@ -32,12 +33,16 @@ final class DisabledConcatenationWithTranslatedStringsRule implements Rule
     /** @var array<class-string, string[]> */
     private array $staticCalls = [];
 
+    /** @var string[] */
+    private array $allowedTranslateConcatenationPatterns;
+
     private NameResolver $nameResolver;
 
     /**
      * @param string[] $translateCalls
+     * @param string[] $allowedTranslateConcatenationPatterns
      */
-    public function __construct(array $translateCalls, NameResolver $nameResolver)
+    public function __construct(array $translateCalls, array $allowedTranslateConcatenationPatterns, NameResolver $nameResolver)
     {
         foreach ($translateCalls as $translateCall) {
             if (!str_contains($translateCall, '::')) {
@@ -65,6 +70,7 @@ final class DisabledConcatenationWithTranslatedStringsRule implements Rule
             }
         }
 
+        $this->allowedTranslateConcatenationPatterns = $allowedTranslateConcatenationPatterns;
         $this->nameResolver = $nameResolver;
     }
 
@@ -83,8 +89,9 @@ final class DisabledConcatenationWithTranslatedStringsRule implements Rule
         }
 
         $errors = [];
-        if ($this->isTranslateCall($node->left, $scope) || $this->isTranslateCall($node->right, $scope)) {
-            $errors[] = RuleErrorBuilder::message('Do not concatenate translated strings.')->tip('Every language has its own word ordering, use variables instead.')->build();
+        if (($this->isTranslateCall($node->left, $scope) && !$this->isAllowedToConcatenate($node->right)) ||
+            ($this->isTranslateCall($node->right, $scope) && !$this->isAllowedToConcatenate($node->left))) {
+            $errors[] = RuleErrorBuilder::message('Do not concatenate translated strings.')->tip('Every language has its own word ordering, use variables in translations instead, e.g. Hello %name%.')->build();
         }
         return $errors;
     }
@@ -126,5 +133,14 @@ final class DisabledConcatenationWithTranslatedStringsRule implements Rule
         }
 
         return false;
+    }
+
+    private function isAllowedToConcatenate(Expr $expr): bool
+    {
+        if (!$expr instanceof String_) {
+            return false;
+        }
+
+        return (bool)preg_match('#' . implode('|', $this->allowedTranslateConcatenationPatterns) . '#', $expr->value);
     }
 }