+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+- Possibility to allow using some defined patterns with translated strings concatenation
+
 ### Fixed
 - Check function call in loop only if it is re-assign of variable